refactor: encapsulate movement logic in TrainMovementManager and remove auto-reverse logic
