+++ src/main/java/letrain/vehicle/rail/impl/Locomotive.java
@@ -264,6 +264,13 @@ public class Locomotive extends Linker implements Tractor {
     }
 
     public void setTargetSpeed(int speed) {
+        if (getTrain() != null && getTrain().getSafetyManager() != null && getTrain().getSafetyManager().isWaitingForBlock()) {
+            if (speed > 0) {
+                log.info("Locomotive {}: Train is waiting for block. Intercepting setTargetSpeed({}) and saving it instead.", id, speed);
+                getTrain().getSafetyManager().onBrakingInitiated(speed);
+                speed = 0;
+            }
+        }
         log.info("Locomotive {}: setTargetSpeed changes from {} to {}", id, this.targetSpeed, speed);
         if (this.targetSpeed == speed) {
             return;
