+++ src/main/java/letrain/vehicle/rail/impl/Train.java
@@ -563,160 +563,11 @@ public class Train implements Trailer<RailTrack>, Renderable, Transportable, Tra
      * @return true if the train moved, false otherwise
      */
     public boolean advance() {
-        if (isLoading()) {
-            log.info("Train {} advance: cannot move because train is loading", id);
-            return false;
-        }
-
-        if (isStalled()) {
-            log.info("Train {} advance: cannot move because train is stalled", id);
-            if (directorLinker != null) {
-                directorLinker.setTargetSpeed(0);
-            }
-            return false;
-        }
-
-        if (model != null) {
-            if (!hasPermissionToMove()) {
-                log.info("Train {} advance: cannot move because hasPermissionToMove is false. Forcing setTargetSpeed(0)", id);
-                if (directorLinker != null) {
-                    directorLinker.setTargetSpeed(0);
-                }
-                return false;
-            }
-        }
-        
-        log.info("Train {} advance: proceeding to moveLinkers", id);
-
-        boolean normalSense = true;
-        if (getDirectorLinker().isReversed()) {
-            normalSense = false;
-        }
-
-        // Save linker directions before attempting to move.
-        // If moveLinkers fails (collision, blocked), we must restore them
-        // so the renderer draws wagons at their correct positions.
-        Map<Linker, letrain.map.Dir> savedDirs = new HashMap<>();
-        Map<Linker, letrain.map.Dir> savedEntryDirs = new HashMap<>();
-        for (Linker l : getLinkers()) {
-            savedDirs.put(l, l.getDir());
-            savedEntryDirs.put(l, l.getEntryDir());
-        }
-
-        setDirPushedLinkers(normalSense);
-        setDirTowedLinkers(normalSense);
-        boolean moved = moveLinkers(normalSense);
-
-        if (!moved || isStalled()) {
-            // Movement failed or train crashed/stalled — restore original
-            // directions. Without this, setDirTowedLinkers leaves wagons
-            // pointing forward, causing the renderer to interpolate them
-            // into the locomotive when the train has actually stopped.
-            // Exception: the first linker's direction was already corrected
-            // by moveLinkers() for the new position. Restoring it would
-            // break the direction on curves after a crash.
-            Linker first = getLinkers().isEmpty() ? null : getLinkers().getFirst();
-            for (Linker l : getLinkers()) {
-                if (isStalled() && l == first)
-                    continue; // skip first linker on crash
-                letrain.map.Dir savedDir = savedDirs.get(l);
-                letrain.map.Dir savedEntry = savedEntryDirs.get(l);
-                if (savedDir != null) {
-                    l.setDir(savedDir);
-                }
-                if (savedEntry != null) {
-                    l.setEntryDir(savedEntry);
-                }
-            }
-        }
-
-        return moved;
+        return movementManager.advance();
     }
 
     public void refreshLinkersDirection() {
-        if (getDirectorLinker() == null || ((Locomotive) getDirectorLinker()).getTrack() == null) {
-            return;
-        }
-        boolean normalSense = true;
-        if (getDirectorLinker().isReversed()) {
-            normalSense = false;
-        }
-        setDirPushedLinkers(normalSense);
-        setDirTowedLinkers(normalSense);
-    }
-
-    private void setDirPushedLinkers(boolean isNormalSense) {
-        Iterator<Linker> iterator;
-        if (!isNormalSense) {
-            iterator = getLinkers().iterator();
-        } else {
-            iterator = getLinkers().descendingIterator();
-        }
-
-        Tractor tractor = getDirectorLinker();
-        while (iterator.hasNext()) {
-            Linker next = iterator.next();
-            if (next == tractor) {
-                break;
-            }
-        }
-        Dir pushDir = ((Locomotive) tractor).getDir();
-        while (iterator.hasNext()) {
-            Linker next = iterator.next();
-            Track nextTrack = next.getTrack();
-            if (pushDir == null) {
-                break;
-            }
-            Dir inverseEntry = pushDir.inverse();
-            next.setEntryDir(inverseEntry);
-            Dir nextDir = nextTrack.getDir(inverseEntry);
-            if (nextDir == null) {
-                break;
-            }
-            next.setDir(nextDir);
-            pushDir = next.getDir();
-        }
-    }
-
-    private void setDirTowedLinkers(boolean isNormalSense) {
-        Iterator<Linker> iterator;
-        if (isNormalSense) {
-            iterator = getLinkers().iterator();
-        } else {
-            iterator = getLinkers().descendingIterator();
-        }
-        Tractor tractor = getDirectorLinker();
-        while (iterator.hasNext()) {
-            Linker next = iterator.next();
-            if (next == tractor) {
-                break;
-            }
-        }
-        Track oldTrack = ((Locomotive) tractor).getTrack();
-        while (iterator.hasNext()) {
-            Linker next = iterator.next();
-            // Find the actual physical connection from oldTrack to the wagon's track,
-            // instead of using geometric Point.locate() which gives diagonal directions
-            // on curves that don't match any track router entry.
-            Dir nextDir = null;
-            Track wagonTrack = next.getTrack();
-            for (Dir conn : oldTrack.getConnections()) {
-                if (oldTrack.getConnected(conn) == wagonTrack) {
-                    nextDir = conn.inverse(); // direction from wagon toward oldTrack
-                    break;
-                }
-            }
-            if (nextDir == null) {
-                break;
-            }
-            next.setDir(nextDir);
-            Dir nextEntry = next.getTrack().getDir(next.getDir());
-            if (nextEntry == null) {
-                break;
-            }
-            next.setEntryDir(nextEntry);
-            oldTrack = next.getTrack();
-        }
+        movementManager.refreshLinkersDirection();
     }
 
     public boolean moveLinkers(boolean isNormalSense) {
